(RE-3710) Update the msi name
With the push for all in one agent packages, we're renaming our puppet
msi. This rename is so that the msi will be more in sync with the rest
of our puppet agent packages. This work is needed to build the windows
installer now that https://tickets.puppetlabs.com/browse/PUP-3844 has
been merged.