(maint) Ensure we fetch necessary vars for gem ship
Prior to this commit, we kept running into issues where centain
variables we not being populated. This was because we never invoke the
`pl:fetch` task, which looks for and sets all remote variables defined
in the build-data repo. New updates to the gem ship task requires this
information, whereas before we didn't. The gem ship will fail unless we
add this task requirement.